refactor(cli): split oversized command modules
